requirements.txt
fastapi==0.115.0
# FastAPI 0.115.0 requires Starlette >=0.37.2,<0.39.0. Pin to avoid runtime incompatibilities.
starlette>=0.37.2,<0.39.0
uvicorn==0.30.6
langchain==0.3.7
langchain-groq==0.2.1
langchain-core==0.3.15
python-dotenv==1.0.1
pydantic==2.9.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.151.0
httpx==0.27.2
chromadb==0.5.15
sentence-transformers==3.3.1
langgraph==0.2.55